+++ tests/EvmAcceptanceTests/helper/DebugHelper.ts
@@ -1,0 +1,4 @@
+import hre from "hardhat";
+
+const logDebug = hre.debug ? console.log.bind(console) : function () {};
+export default logDebug;
